refactor: moved project DTOs to dto package
