--- src/main/java/com/pawelnu/projectmanager/endpoints/project/ProjectSimpleDTO.java
@@ -1,3 +1,0 @@
-package com.pawelnu.projectmanager.endpoints.project;
-
-public class ProjectSimpleDTO {}
