+++ src/main/java/com/pawelnu/projectmanager/endpoints/project/dto/ProjectCreateRequestDTO.java
@@ -1,4 +1,4 @@
-package com.pawelnu.projectmanager.endpoints.project;
+package com.pawelnu.projectmanager.endpoints.project.dto;
 
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
