+++ src/main/java/com/pawelnu/projectmanager/endpoints/project/dto/ProjectSimpleDTO.java
@@ -1,0 +1,3 @@
+package com.pawelnu.projectmanager.endpoints.project.dto;
+
+public class ProjectSimpleDTO {}
